REFINEMENT: Fix return value kind in check_refinement_criterion().
The meta block field %refine is 4 byte integer. Make the value returned
by check_refinement_criterion() of the same type.

+++ src/refinement.F90
@@ -246,7 +246,7 @@ module refinement
 
 ! return variable
 !
-    integer(kind=1)                       :: criterion
+    integer(kind=4)                       :: criterion
 
 ! local variables
 !
